+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-1230](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1230)
 - Added the option for Layers on the Analysis Page Map
   [GEOD-1225](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1225)
+- Added static banners
+  [GEOD-1203](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1203)
 
 ### Changed
 - The Search Page Filter Icons reflect the dropdown selection